Addition:SpringBootProject

+++ sprintBoot_Demo/JournalApp/journalapp/src/main/java/com/sgtsoftsol/journalapp/service/JournalEntryService.java
@@ -7,6 +7,7 @@ import java.util.Optional;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.sgtsoftsol.journalapp.entity.JournalEntry;
 import com.sgtsoftsol.journalapp.entity.User;
@@ -22,37 +23,56 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
-    // JournalEntryService(JournalappApplication journalappApplication, JournalEntryController journalEntryController) {
-    //     this.journalappApplication = journalappApplication;
-    //     this.journalEntryController = journalEntryController;
+    // JournalEntryService(JournalappApplication journalappApplication,
+    // JournalEntryController journalEntryController) {
+    // this.journalappApplication = journalappApplication;
+    // this.journalEntryController = journalEntryController;
     // }
 
-    public void saveEntry(JournalEntry journalEntry, String username){
-        User user = userService.findByUserName(username);
-        journalEntry.setDate(LocalDateTime.now());
-        JournalEntry saved = journalEntryRepository.save(journalEntry);
-        user.getJournalEntries().add(saved);
-        userService.saveEntry(user);
+    @Transactional
+    public void saveEntry(JournalEntry journalEntry, String username) {
+        try {
+            User user = userService.findByUserName(username);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            userService.saveUser(user);
+
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while saving the entry", e);
+        }
     }
 
-    public void saveEntry(JournalEntry journalEntry){
+    public void saveEntry(JournalEntry journalEntry) {
         journalEntryRepository.save(journalEntry);
     }
 
-    public List<JournalEntry> getAll(){
+    public List<JournalEntry> getAll() {
         return journalEntryRepository.findAll();
     }
 
-    public Optional<JournalEntry> findById(ObjectId id){
+    public Optional<JournalEntry> findById(ObjectId id) {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id,String username){
-        User user = userService.findByUserName(username);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String username) {
+        boolean removed=false;
+        try {
+            User user = userService.findByUserName(username);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry.",e);
+        }
+
+        return removed;
     }
 
-    
 }
